Fri, Mar 3, 2023, 3:14 PM +08:00

+++ readability/readability.c
@@ -34,7 +34,7 @@ int main(void)
 
  */
     //Compute Coleman-Liau index
-    int index = (0.0588 * ((letters / words) * 100)) - (0.296 * ((sentences / words) * 100)) - 15.8;
+    float index = (float) (0.0588 * ((letters / words) * 100)) - (0.296 * ((sentences / words) * 100)) - 15.8;
 
     if (index < 1)
     {
@@ -46,7 +46,7 @@ int main(void)
     }
     else //Grade 1 to 15
     {
-        printf("Grade %i\n", index);
+        printf("Grade %i\n", (int) index);
     }
 }
